Fix CI by adding missing catkin-pkg-modules
* Implement abi parameter in build_task_generator
* Move call_abi_checker to the only file using it
* Missing python3-catkin-pkg-modules to run testing
