+++ .travis.yml
@@ -197,8 +197,10 @@ matrix:
         - docker
       env: JOB_TYPE=external_prerelease
       before_script:
+        - sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros.list'
+        - *rosgpgkeyadd
         - sudo apt-get update
-        - sudo apt-get install -y python3-pip python3-setuptools
+        - sudo apt-get install -y python3-pip python3-setuptools python3-catkin-pkg-modules
         - sudo pip3 install vcstool
         - python setup.py install
         - mkdir job && cd job
@@ -285,6 +287,7 @@ matrix:
         - sudo apt-get update -qq
         - sudo apt-get install dpkg -y  # necessary for catkin-pkg to be installable
         - sudo apt-get install ros-kinetic-catkin -y
+        - sudo apt-get install python3-catkin-pkg-modules -y
         - python setup.py install
         - mkdir job && cd job
         - ln -s .. ros_buildfarm
@@ -300,8 +303,10 @@ matrix:
         - docker
       env: JOB_TYPE=external_prerelease
       before_script:
+        - sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros.list'
+        - *rosgpgkeyadd
         - sudo apt-get update
-        - sudo apt-get install -y python3-pip python3-setuptools
+        - sudo apt-get install -y python3-pip python3-setuptools python3-catkin-pkg-modules
         - sudo pip3 install vcstool
         - python setup.py install
         - mkdir job && cd job
@@ -388,6 +393,7 @@ matrix:
         - sudo apt-get update -qq
         - sudo apt-get install dpkg -y  # necessary for catkin-pkg to be installable
         - sudo apt-get install ros-kinetic-catkin -y
+        - sudo apt-get install python3-catkin-pkg-modules -y
         - python setup.py install
         - mkdir job && cd job
         - ln -s .. ros_buildfarm
@@ -403,8 +409,10 @@ matrix:
         - docker
       env: JOB_TYPE=external_prerelease
       before_script:
+        - sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros.list'
+        - *rosgpgkeyadd
         - sudo apt-get update
-        - sudo apt-get install -y python3-pip  python3-setuptools
+        - sudo apt-get install -y python3-pip  python3-setuptools python3-catkin-pkg-modules
         - sudo pip3 install vcstool
         - python setup.py install
         - mkdir job && cd job
@@ -485,8 +493,10 @@ matrix:
         - docker
       env: JOB_TYPE=prerelease UNDERLAY_REPOSITORY_NAMES="roscpp_core" OVERLAY_PACKAGE_NAMES=roscpp
       before_script:
+        - sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros.list'
+        - *rosgpgkeyadd
         - sudo apt-get update
-        - sudo apt-get install -y python3-pip python3-setuptools
+        - sudo apt-get install -y python3-pip python3-setuptools python3-catkin-pkg-modules
         - sudo pip3 install EmPy
         # install catkin for test results
         - sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros.list'
@@ -509,8 +519,10 @@ matrix:
         - docker
       env: JOB_TYPE=external_prerelease
       before_script:
+        - sudo sh -c 'echo "deb http://packages.ros.org/ros/ubuntu $(lsb_release -sc) main" > /etc/apt/sources.list.d/ros.list'
+        - *rosgpgkeyadd
         - sudo apt-get update
-        - sudo apt-get install -y python3-pip python3-setuptools
+        - sudo apt-get install -y python3-pip python3-setuptools python3-catkin-pkg-modules
         - sudo pip3 install EmPy vcstool
         - python setup.py install
         - mkdir job && cd job
